refractor purge channel route
- tested
